fix(renaming): change names components
