+++ src/App.js
@@ -1,7 +1,7 @@
 import { Routes, Route } from "react-router-dom";
 import Layout from './components/Layout/Layout.jsx'
 import DetailCharacter from "./pages/Home/DetailCharacter/DetailCharacter.jsx"; 
-import Filters from "./components/Filters/Filters.jsx";
+import Home from "./components/Home/Home.jsx";
 
 
 function App() {
@@ -9,7 +9,7 @@ function App() {
     <>
     <Layout/>
       <Routes>
-        <Route path="/" element={<Filters />} />
+        <Route path="/" element={<Home />} />
         <Route path="character/:characterId" element={<DetailCharacter/>}/>
       </Routes>
     </>
