+++ src/pages/Home/ListCharacters/ListCharacters.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.module.scss';
+@import '../../../styles/variables.module.scss';
 
 .home-container{
     display: grid;
